remove trait upcasting experiment macro
