+++ src/lib.rs
@@ -3,9 +3,9 @@
 // Fortunately, this feature will be stabilized soon at this point of time(Thanks for Rust teams),
 // and I look forward to the day when this project finished and I can remove this annotation without concern.
 // See issue: https://github.com/rust-lang/rust/issues/65991
-// todo: remove this
-#![feature(trait_upcasting)]
-#![allow(incomplete_features)]
+// removed at 2023-12-28
+// #![feature(trait_upcasting)]
+// #![allow(incomplete_features)]
 
 // todo: look closely at all function annotated by #[inline], consider using it properly.
 
@@ -44,8 +44,8 @@ pub mod atn_deserializer;
 pub mod atn_deserialize_option;
 pub mod interval_set;
 pub mod transition;
-pub mod misc;
 pub mod semantic_context;
+pub mod misc;
 
 /// use [Nth] to present a target position
 pub type Nth = usize;
