+++ tests/arishem_lexer_test.rs
@@ -1,7 +1,6 @@
-use std::cell::RefCell;
-use std::rc::Rc;
-use std::sync::{Arc, RwLock};
 use lazy_static::lazy_static;
+
+use std::sync::{Arc, RwLock};
 use antlr4rs::atn::ATN;
 use antlr4rs::atn_deserializer::ATNDeserializer;
 use antlr4rs::dfa::DFA;
