+++ tests/tree_visitor_test.rs
@@ -1,14 +1,11 @@
-#![feature(trait_upcasting)]
-#![allow(incomplete_features)]
+// #![feature(trait_upcasting)]
+// #![allow(incomplete_features)]
 
 use std::any::Any;
 use std::borrow::Cow;
-use std::fmt::Debug;
 use std::rc::Rc;
 use antlr4rs::parser_rule_context::BaseParserRuleContext;
-use antlr4rs::rule_context::RuleContext;
-use antlr4rs::token::Token;
-use antlr4rs::tree::{ErrorNode, ParseTree, ParseTreeVisitor, RuleNode, SyntaxTree, TerminalNode, Tree};
+use antlr4rs::tree::{ParseTree, ParseTreeVisitor, SyntaxTree, Tree};
 use antlr4rs::value::Val;
 use antlr4rs::value::Val::{Nil, StrSRef};
 
@@ -86,7 +83,7 @@ impl MyParseTreeVisitorProxy {
     }
 }
 
-#[derive(Debug)]
+/*#[derive(Debug)]
 struct AnyAnyAny {
     name: String,
     age: usize,
@@ -96,7 +93,7 @@ impl AnyAnyAny {
     fn new(name: String, age: usize) -> Self {
         Self { name, age }
     }
-}
+}*/
 
 struct IamVisitor {}
